requirements.txt
Flask==1.1.2
Flask-WTF==0.14.3
wtforms-validators==1.0.0
passlib==1.7.4
pillow==8.1.0
PyMySQL==1.0.2
pathlib==1.0.1
email-validator==1.1.2
flask-paginate==0.8.1